Rename folders to look nice in github
