+++ 01-ancient-egyptian-multiplication/multiply.js
