+++ 01-ancient-egyptian-multiplication/multiply.spec.js
