+++ 02-linear-search/linearSearch.js
