+++ 04-greedy-algorithm/minimumCoins.js
