+++ 04-greedy-algorithm/minimumCoins.spec.js
